+++ eq/server/config.h
@@ -166,9 +166,7 @@ public:
 
     virtual void restore();
 
-    EventOCommand sendError( const uint32_t type,
-                             const uint128_t& originator,
-                             const uint32_t error );
+    EventOCommand sendError( const uint32_t type, const Error& error );
 
     /** Return the initID, used for late initialization  */
     uint128_t getInitID(){ return _initID; }
